FindDerivativesButton: move to @labkey/components/entities
