+++ packages/components/src/entities/FindDerivativesButton.spec.tsx
@@ -4,15 +4,15 @@ import { Filter } from '@labkey/api';
 
 import { fromJS } from 'immutable';
 
-import { makeTestQueryModel } from '../../../public/QueryModel/testUtils';
-import { SchemaQuery } from '../../../public/SchemaQuery';
-import { mountWithAppServerContext } from '../../testHelpers';
-import { QueryInfo } from '../../../public/QueryInfo';
-import { ViewInfo } from '../../ViewInfo';
+import { makeTestQueryModel } from '../public/QueryModel/testUtils';
+import { SchemaQuery } from '../public/SchemaQuery';
+import { mountWithAppServerContext } from '../internal/testHelpers';
+import { QueryInfo } from '../public/QueryInfo';
+import { ViewInfo } from '../internal/ViewInfo';
 
-import { DisableableMenuItem } from '../samples/DisableableMenuItem';
+import { DisableableMenuItem } from '../internal/components/samples/DisableableMenuItem';
 
-import { DataClassDataType, SampleTypeDataType } from './constants';
+import { DataClassDataType, SampleTypeDataType } from '../internal/components/entities/constants';
 
 import { FindDerivativesMenuItem, getFieldFilter, getSessionSearchFilterProps } from './FindDerivativesButton';
